Refactor: Extract logic into custom hooks
This refactoring extracts the drawing, keyboard controls, and SVG export logic from App.tsx into three separate custom hooks:

- `useDrawing`: Handles all mouse-related drawing logic.
- `useKeyboardControls`: Manages keyboard events for deleting shapes.
- `useSvgExport`: Contains the logic for exporting the SVG canvas.

This change improves code organization and readability by separating concerns, making the App.tsx component cleaner and easier to maintain.
